Generalized apply_binary_op & apply_unary_op
